requirements.txt
streamlit>=1.0
mysql-connector-python
pandas
numpy
